Don't allow apps in a web browser to quit
