+++ internal/driver/glfw/menu_notweb.go
@@ -1,0 +1,35 @@
+//go:build !js && !wasm && !test_web_driver
+// +build !js,!wasm,!test_web_driver
+
+package glfw
+
+import "fyne.io/fyne/v2"
+
+func addMissingQuitForMainMenu(menus *fyne.MainMenu, w *window) *fyne.MainMenu {
+	var lastItem *fyne.MenuItem
+	if len(menus.Items[0].Items) > 0 {
+		lastItem = menus.Items[0].Items[len(menus.Items[0].Items)-1]
+		if lastItem.Label == "Quit" {
+			lastItem.IsQuit = true
+		}
+	}
+	if lastItem == nil || !lastItem.IsQuit { // make sure the first menu always has a quit option
+		quitItem := fyne.NewMenuItem("Quit", nil)
+		quitItem.IsQuit = true
+		menus.Items[0].Items = append(menus.Items[0].Items, fyne.NewMenuItemSeparator(), quitItem)
+	}
+	for _, item := range menus.Items[0].Items {
+		if item.IsQuit && item.Action == nil {
+			item.Action = func() {
+				for _, win := range w.driver.AllWindows() {
+					if glWin, ok := win.(*window); ok {
+						glWin.closed(glWin.view())
+					} else {
+						win.Close() // for test windows
+					}
+				}
+			}
+		}
+	}
+	return menus
+}
