+++ internal/driver/glfw/menu_web.go
@@ -1,0 +1,11 @@
+//go:build js || wasm || test_web_driver
+// +build js wasm test_web_driver
+
+package glfw
+
+import "fyne.io/fyne/v2"
+
+func addMissingQuitForMainMenu(menus *fyne.MainMenu, w *window) *fyne.MainMenu {
+	// no-op for a web browser
+	return menus
+}
